requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
google-generativeai==0.8.3
requests==2.31.0
PyPDF2==3.0.1
gunicorn==21.2.0